+++ client/CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - Changed Add Row button styling in Advanced Search ([#642](https://github.com/project-lux/lux-frontend/issues/642)).
+- Changed the input row and relationship row styling in advanced search ([#643](https://github.com/project-lux/lux-frontend/issues/643)).
 
 ### Removed