+++ client/CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 
 - Changed Add Row button styling in Advanced Search ([#642](https://github.com/project-lux/lux-frontend/issues/642)).
 - Changed the input row and relationship row styling in advanced search ([#643](https://github.com/project-lux/lux-frontend/issues/643)).
+- Changed the multi-field row styling in advanced search ([#644](https://github.com/project-lux/lux-frontend/issues/644)).
 
 ### Removed